Добавлена Policy для запрета удаления чужих заданий. Удалено ненужное поле из сущности "Задание" в бд.

+++ app/Http/Controllers/TaskController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\Task\StoreRequest;
 use App\Models\Task;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
 use Illuminate\Foundation\Application;
@@ -66,9 +67,11 @@ class TaskController extends Controller
 
     /**
      * Remove the specified resource from storage.
+     * @throws AuthorizationException
      */
     public function destroy(Task $task): RedirectResponse
     {
+        $this->authorize('destroy', $task);
         $task->delete();
         return redirect()->route('tasks.index');
     }
